ui/test: add a 2nd E2E test for volume creation
Now check that we can also create a sparseLoopDevice Volume

Refs: #1429
